third_party/WebKit/LayoutTests/gamepad/gamepad-api.html

<!DOCTYPE html>
<body>
<script src="../resources/js-test.js"></script>
<script>
    shouldBeDefined("navigator.webkitGetGamepads");
    shouldBeDefined("navigator.getGamepads");
</script>
<p>Make sure the main polling access point exists on navigator.</p>
</body>
